Raise a werkzeug exception when the user is over their rate limit
Returning a body and status code means that this response will be
sent directly to the client as text/plain.
Using an exception allows us to use flask's request error handling to
format the error using the upstream app's configuration so that we can,
for example, format the error as JSON for API responses and HTML for
website responses.