Always return rating if it exists, only do 1 query for ratings
Previously we would do 1 query to get an entity and then another
to get the rating (and then only 1 query to get an entity and its rating
if we joined to the redirects table!).
This simplifies the queries to always join to the meta table if it
exists.
Because we always join to this table anyway, remove the rating includes
(it was spelt wrong anyway and never returned data) and always return it