Group release groups by rg and rg_meta fields to prevent duplicates
Because we join against the release table to get the release label, the
count query will include each release, in the case that there are
multiple releases for that releasegroup.
Group by release_group.* and release_group_meta.* (because we sort on
it).
The SQLAlchemy ORM was already deduplicating these items, so there was a
mismatch in the count and number of returned items. Now these values are
consistent.

Add a contains_eager load for release_group_primary_time to prevent n+1
queries

Only filter by release type if a type is provided. This allows for a
query to get all release groups by providing an empty type list.