Remove decorator override for ratelimit
The ratelimit keys in the cache, include a global or scope prefix in the key name.
At the decorator level, there is no prefix, and hence it would use the global prefix
and end up overriding global ratelimit records for the requests made by the user in
the current window. Thinking more of it, I don't think we need decorator level overrides.
Scope isolation should be sufficient.